general clean-up + corrections
Tuesday backend work has been changed to remove discrepancies between all classes. Corrections have been made (missing getters and setters, misnamed attributes, missing constructors, redundant [?] constructors, etc. Spacing and position has been coordinated between all classes. Minimal comments to facilitate browsing models have been added.

+++ src/main/java/com/example/bigbowlxp/repository/BowlingLaneRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BowlingLaneRepository extends JpaRepository<BowlingLane, Integer> {
+public interface BowlingLaneRepository extends JpaRepository<BowlingLane, Long> {
 }
