+++ src/main/java/com/example/bigbowlxp/repository/HockeyBookingRepository.java
@@ -3,5 +3,5 @@ package com.example.bigbowlxp.repository;
 import com.example.bigbowlxp.model.HockeyBooking;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface HockeyBookingRepository extends JpaRepository<HockeyBooking, Integer> {
+public interface HockeyBookingRepository extends JpaRepository<HockeyBooking, Long> {
 }
