+++ src/main/java/com/example/bigbowlxp/repository/HockeyTableRepository.java
@@ -3,5 +3,5 @@ package com.example.bigbowlxp.repository;
 import com.example.bigbowlxp.model.HockeyTable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface HockeyTableRepository extends JpaRepository<HockeyTable, Integer> {
+public interface HockeyTableRepository extends JpaRepository<HockeyTable, Long> {
 }
